refactor(commands): extract runtimes and response writer
